+++ clang/lib/Analysis/ThreadSafety.cpp
@@ -2081,7 +2081,8 @@ void BuildLockset::VisitUnaryOperator(const UnaryOperator *UO) {
       checkAccess(UO->getSubExpr(), AK_Written);
       break;
     case UO_AddrOf:
-      checkAccess(UO->getSubExpr(), AK_Read, POK_AddressOf);
+      if (Analyzer->Handler.issueBetaWarnings())
+        checkAccess(UO->getSubExpr(), AK_Read, POK_AddressOf);
       break;
     default:
       break;